refactor: consolidate logfire into a single module
Merge flags.ts, logfire.ts, and providers/logfire.ts into src/logfire.ts.
Registry imports wrapWithLogfire from there. Fewer files to maintain.
